libssh: reverting DrvCloudTunnel.cpp . bugref:10539
svn:sync-xref-src-repo-rev: r159713

+++ src/VBox/Devices/Network/DrvCloudTunnel.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvCloudTunnel.cpp 159711 2023-10-27 13:37:10Z alexander.rudnev@oracle.com $ */
+/* $Id: DrvCloudTunnel.cpp 159713 2023-10-27 13:47:17Z alexander.rudnev@oracle.com $ */
 /** @file
  * DrvCloudTunnel - Cloud tunnel network transport driver
  *
@@ -754,7 +754,7 @@ static int drvCloudTunnelReceiveCallback(ssh_session session, ssh_channel channe
     return len;
 }
 
-static int channelWriteWontblockCallback(ssh_session, ssh_channel, unsigned int, void *)
+static int channelWriteWontblockCallback(ssh_session, ssh_channel, size_t, void *)
 {
     return 0;
 }
